Move action classes to own package
